refactored the AST_ProcedureMap and AST_ModuleStack moduled
the procedure map no longer owns identifiers but only keeps a new clone of module and procedure names ans string. itbis freed when the map od cleared. identifiers for module and procedure declarations are discarded as soon as theyre used to create the ast form of the module and procedure.